pyproject.toml
[project]
name = "pi-pico"
version = "0.1.0"
description = "這是樹梅派和pico整合的專案"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.1.0",
    "paho-mqtt>=2.1.0",
    "streamlit>=1.52.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
]
